Create Subpalindrome.py
A palindrome is a string that reads the same from left and from the right. For example mom, tacocat and malayalam are palindromes as are any single character strings. Given a string, determine the number of its substrings that are palindromes. For example the string s = tacocat, palindrome substrings are (t,a,c,o,c,a,t,coc,acoca,tacocat). Totally there are 10 palindrome substrings. So you want to return the output as 10.