Create StringVowels.py
Input string should be splitted with the vowel characters 'a','e','i','o','u'. Identify the splitting position begins with 1. The output string should be processed based on the given condition. 1. If the string position is odd then the string should be printed or appended as it is with out any change. 2. If the string position is even then the string should be printed or appended in reverse order.